Circumventing a bug in mime-util library (http://sourceforge.net/tracker/?func=detail&aid=3007610&group_id=205064&atid=992132).
mime-util leaks threads when called repeatedly (specifically, OpendesktopMimeDetector class). Using a singleton causes an IllegalArgumentException in OpendesktopMimeDetector. The workaround for now is to use: a singleton, a one permit semaphore, and retrying if an IllegalArgumentException is caught. This seems to perform faster than calling repeatedly and invoking an explicit delete.